Suggest using verbose option (-v) for more explicit error messages.
Instead of editing the source (which will be overwritten every time the package is updated), use the `-v` (verbose) flag instead.  This provides a diff-like output along with context of the failed check and a stack-trace.

The output is especially helpful for arrays and objects that aren't returning quite right as it shows what keys or properties are missing.